+++ webpack.config.js
@@ -1,9 +1,7 @@
 const path = require('path');
 const webpack = require('webpack');
 
-const TerserPlugin = require('terser-webpack-plugin');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 const ScratchWebpackConfigBuilder = require('scratch-webpack-configuration');
 
@@ -12,13 +10,6 @@ const baseConfig = new ScratchWebpackConfigBuilder({
 })
     .enableDevServer(process.env.PORT || 8361)
     .merge({
-        optimization: {
-            minimizer: [
-                new TerserPlugin({
-                    include: /\.min\.js$/
-                })
-            ]
-        },
         resolve: {
             fallback: {
                 Buffer: require.resolve('buffer/')
@@ -41,19 +32,7 @@ const webConfig = baseConfig.clone()
                 name: 'ScratchRender'
             }
         }
-    })
-    .addPlugin(new HtmlWebpackPlugin({
-        chunks: 'all',
-        filename: 'index.html',
-        template: 'test/integration/index.ejs',
-        scriptLoading: 'blocking'
-    }))
-    .addPlugin(new HtmlWebpackPlugin({
-        chunks: 'all',
-        filename: 'cpu-render.html',
-        template: 'test/integration/cpu-render.ejs',
-        scriptLoading: 'blocking'
-    }));
+    });
 
 const playgroundConfig = baseConfig.clone()
     .setTarget('browserslist')